message sent at

+++ FlashTape/FriendTableViewCell.m
@@ -48,7 +48,7 @@
     if (currentUserPostsCount != 0) {
         self.saveButton.hidden = NO;
         self.storyVideosCountLabel.hidden = NO;
-        self.storyVideosCountLabel.text = [NSString stringWithFormat:@"%lu",currentUserPostsCount];
+        self.storyVideosCountLabel.text = [NSString stringWithFormat:@"%lu",(long)currentUserPostsCount];
         self.storyVideosCountLabel.layer.borderWidth = 1;
         self.storyVideosCountLabel.layer.borderColor = [UIColor whiteColor].CGColor;
         self.storyVideosCountLabel.layer.cornerRadius = self.storyVideosCountLabel.frame.size.height / 2;
